Fix DOCKER_CROSS being overwritten
Not exactly sure why, but this line;

   build: DOCKER_CROSS ?= false

Always overwrote `DOCKER_CROSS` when running `make cross`.
Perhaps because it is set in `cross: DOCKER_CROSS := true`,
and in a different scope? May also be dependent on the
version of `make` in use.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>
(cherry picked from commit 934b3a3841fb47649ccf93257ed909cbd0f1caed)
Signed-off-by: Sebastiaan van Stijn <github@gone.nl>